% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_status.R
\name{build_pdb}
\alias{build_pdb}
\title{Formats data received from PDB for merging with stocksmart data}
\usage{
build_pdb(stocknames, fnames, year)
}
\arguments{
\item{stocknames}{a data frame with stock names and their abbreviations, used to replace the the PDB stock \code{ABBR} with a stocksmart \code{Entity.Name} in the output}

\item{fnames}{a character vector of PDB file names to read in}

\item{year}{a numeric value representing the assessment year, which will be added to the output data frame}
}
\value{
a tibble
}
\description{
This function reads in stock status files from PDB and returns the data formatted for merging with stocksmart data.
}
