% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_biomass.r
\name{create_aggregate_biomass}
\alias{create_aggregate_biomass}
\title{Calculates aggregate_biomass data set for automated workflow}
\usage{
create_aggregate_biomass(input_path_survey, input_path_species)
}
\arguments{
\item{input_path_survey}{Character string. Full path to the survdat data pull rds file}

\item{input_path_species}{Character string. Full path to the species list data pull rds file}
}
\value{
list
\item{aggregate_biomass}{The \code{ecodata::aggregate_biomass} data frame}
\item{aggregate_biomass_species}{Stratified mean for each species/Season at EPU level that make up the aggregate}
}
\description{
This uses the survdat data pull from the survey package and creates EPU
and shelfwide indicators. It is formatted exactly like the ecodata data object
}
\examples{
\dontrun{
# create the ecodata::aggregate_biomass indicator
create_aggregate_biomass(input_path_survey = "path/to/survdatData.rds",
                         input_path_species = "path/to/species.rds")

}


}
