% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_biomass.r
\name{create_aggregate_biomass_epu}
\alias{create_aggregate_biomass_epu}
\title{Calculates Aggregate biomass data set by epu for automated workflow}
\usage{
create_aggregate_biomass_epu(inputPathSurvey, inputPathSpecies, end.year)
}
\arguments{
\item{inputPathSurvey}{Input survey data full file path}

\item{inputPathSpecies}{Input species list full file path}

\item{end.year}{End year for the data}
}
\value{
data frame
}
\description{
Calculates Aggregate biomass data set by epu for automated workflow
}
