% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_biomass.r
\name{create_aggregate_biomass_shelfwide}
\alias{create_aggregate_biomass_shelfwide}
\title{Calculates shelfwide aggregate biomass data set for automated workflow}
\usage{
create_aggregate_biomass_shelfwide(inputPathSurvey, inputPathSpecies, end.year)
}
\arguments{
\item{inputPathSurvey}{Input survey data full file path}

\item{inputPathSpecies}{Input species list full file path}

\item{end.year}{End year for the data}
}
\value{
data frame
}
\description{
Calculates shelfwide aggregate biomass data set for automated workflow
}
\examples{
\dontrun{
# create the ecodata::aggregate_biomass shelfwide component for 2025
create_aggregate_biomass_shelfwide(inputSurvey = "path/to/survdatData.rds"),
                         inputSpecies = "path/to/species.rds"),
                         end.year = 2024)

}

}
