% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_n.R
\name{create_exp_n}
\alias{create_exp_n}
\title{Calculates expected number of species for automated workflow}
\usage{
create_exp_n(input_path_albatross, input_path_bigelow)
}
\arguments{
\item{input_path_albatross}{Character string. Full path to the Albatross data pull rds file}

\item{input_path_bigelow}{Character string. Full path to the Bigelow data pull rds file}
}
\value{
ESn.epu, ecodata::exp_n data frame
}
\description{
This uses the survdat data pull from the survey package.
It is formatted exactly like the ecodata data object
This calculates the expected number of species per tow in the NEFSC Bottom Trawl Survey for Fall and Spring
}
\examples{
\dontrun{
# create the ecodata::exp_n indicator
create_exp_n(input_path_albatros = "path/to/albatross.rds",
             input_path_bigelow = "path/to/bigelow.rds")

}

}
