% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatwave.R
\name{create_heatwave_surface}
\alias{create_heatwave_surface}
\title{Calculates surface temperature portion of heatwave data set for automated workflow}
\usage{
create_heatwave_surface(inputPathGB, inputPathGOM, inputPathMAB)
}
\arguments{
\item{inputPathGB}{Character string. Full path to the GB GLORYS input file from Kim Hyde}

\item{inputPathGOM}{Character string. Full path to the GOM GLORYS input file from Kim Hyde}

\item{inputPathMAB}{Character string. Full path to the MAB GLORYS input file from Kim Hyde}
}
\value{
surface temperature portion of ecodata::heatwave data frame
}
\description{
This uses a static input file from Kim Hyde
It is formatted exactly like the ecodata data object
}
\examples{
\dontrun{
# create the surface temperature portion of the ecodata::heatwave indicator for 2025
create_heatwave_surface(inputPathGB = "path/to/input/GBdata.csv",
                        inputPathGOM = "path/to/input/GOMdata.csv",
                        inputPathMAB = "path/to/input/MABdata.csv")

}


}
