% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatwave_year.R
\name{create_heatwave_year}
\alias{create_heatwave_year}
\title{Calculates the heatwave_year indicator}
\usage{
create_heatwave_year(
  input_path_gb_bot,
  input_path_gom_bot,
  input_path_mab_bot,
  input_path_gb_surf,
  input_path_gom_surf,
  input_path_mab_surf
)
}
\arguments{
\item{input_path_gb_bot}{Character string. Full path to the GB GLORYS input file from Joe Caracappa}

\item{input_path_gom_bot}{Character string. Full path to the GOM GLORYS input file from Joe Caracappa}

\item{input_path_mab_bot}{Character string. Full path to the MAB GLORYS input file from Joe Caracappa}

\item{input_path_gb_surf}{Character string. Full path to the GB OISST input file from Kim Hyde}

\item{input_path_gom_surf}{Character string. Full path to the GOM OISST input file from Kim Hyde}

\item{input_path_mab_surf}{Character string. Full path to the MAB OISST input file from Kim Hyde}
}
\value{
ecodata::heatwave data frame
}
\description{
Combines the surface SST, and the bottom temperature components
into a single indicator
}
\examples{
\dontrun{
# create the ecodata::heatwave_year indicator
create_heatwave_year(input_path_gb_bot = "path/to/input/GBdata.csv",
                        input_path_gom_bot = "path/to/input/GOMdata.csv",
                        input_path_mab_bot = "path/to/input/MABdata.csv",
                        input_path_gb_surf = "path/to/input/GBdata.csv",
                        input_path_gom_surf = "path/to/input/GOMdata.csv",
                        input_path_mab_surf = "path/to/input/MABdata.csv")

}

}
