% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatwave_year_bottom.R
\name{create_heatwave_year_bottom}
\alias{create_heatwave_year_bottom}
\title{Calculates bottom temperature portion of heatwave_year data set for automated workflow}
\usage{
create_heatwave_year_bottom(inputPathGB, inputPathGOM, inputPathMAB)
}
\arguments{
\item{inputPathGB}{Character string. Full path to the GB GLORYS input file from Joe Caracappa}

\item{inputPathGOM}{Character string. Full path to the GOM GLORYS input file from Joe Caracappa}

\item{inputPathMAB}{Character string. Full path to the MAB GLORYS input file from Joe Caracappa}
}
\value{
bottom temperature portion of ecodata::heatwave_year data frame
}
\description{
This uses a static input file from Joe Caracappa
It is formatted exactly like the ecodata data object
}
\examples{
\dontrun{
# create the bottom temperature portion of the ecodata::heatwave_year indicator for 2025
create_heatwave_year_bottom(inputPathGB = "path/to/input/GBdata.csv",
                       inputPathGOM = "path/to/input/GOMdata.csv",
                       inputPathMAB = "path/to/input/MABdata.csv")

}


}
