% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_inshore_survey.R
\name{create_mass_inshore_survey}
\alias{create_mass_inshore_survey}
\title{Function to process Massachusetts inshore bottom trawl survey for automated workflow}
\usage{
create_mass_inshore_survey(inputPathMassSurvey, input_path_species)
}
\arguments{
\item{inputPathMassSurvey}{Character string. Full path to the mass inshore data pull rds file created by workflow_pull_survey_data}

\item{input_path_species}{Character string. Full path to the species list data pull rds file}
}
\value{
ecodata::mass_inshore_survey data frame
}
\description{
Data include aggregated time series of inshore fishery-independent trawl survey data from Massachusetts waters.
MA inshore surveys have been performed biannually in the spring and fall since 1978.
}
\examples{
\dontrun{
# create the ecodata::mass_inshore_survey indicator
create_mass_inshore_survey(inputPathMassSurvey <- here::here("mass_inshore.rds"),
                         input_path_species <- "/home/<user>/EDAB_Datasets/SOE_species_list_24.rds")

}


}
