% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_status.R
\name{create_pdb}
\alias{create_pdb}
\title{Creates stocksmart-like object from files received from PDB}
\usage{
create_pdb(dirfname)
}
\arguments{
\item{dirfname}{A character string representing the directory where the stock status files are stored}
}
\value{
a stocksmart-like tibble
}
\description{
This function reads in a stock status file from PDB and returns a stocksmart-like object. It was created based on the files for spring 2024 MT assessments, which were sent via email from Kristan Blackhart to Scott. Dan Hennen created the files.
This function is used by \code{build_pdb()}.
}
