% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_condition.R
\name{create_species_condition}
\alias{create_species_condition}
\title{Create species condition indicator}
\usage{
create_species_condition(
  inputPath,
  inputpathLW,
  inputpathSpecies,
  by_EPU = TRUE,
  by_sex = FALSE,
  length_break = NULL,
  more_than_20_years = TRUE,
  record_outliers = FALSE,
  output = "soe"
)
}
\arguments{
\item{inputPath}{Character string. Full path to the condition data pull rds file.}

\item{inputpathLW}{Character string. Full path to the LWparams rda file in '//nefscdata/EDAB_Resources'.}

\item{inputpathSpecies}{Character string. Full path to the species.codes rda file in '//nefscdata/EDAB_Resources'.}
}
\value{
condition, ecodata::condition data frame
}
\description{
This function calculates mean relative condition calculated from the NEFSC bottom trawl survey
Methods derived from Laurel Smith (https://github.com/Laurels1/Condition/blob/master/R/RelConditionEPU.R)
}
\examples{
\dontrun{
#create the ecodata::condition indicator
species_condition(
 inputpathLW = "path/to/LWparams.rda",
 inputpathSpecies = "path/to/species.codes.rda",
 by_EPU = TRUE,
 by_sex = FALSE,
 length_break = NULL,
 more_than_20_years = TRUE,
 record_outliers = FALSE,
 output = "soe")

}

}
