% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_status.R
\name{create_stock_status}
\alias{create_stock_status}
\title{Creates the stock status data frame for the SOE}
\usage{
create_stock_status(input_path_decoder)
}
\arguments{
\item{input_path_decoder}{a data frame that matches the stock names with a code abbreviation to use in plotting, typically read from a CSV file. If set to \code{FALSE}, the function will return a data frame in a less processed form, like Sarah used to provide in the \code{assess.csv} file.}
}
\value{
a tibble
}
\description{
This function creates a stock status data frame that is formatted for inclusion in the \code{ecodata} R package.
}
