% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_dates.R
\name{create_trans_dates}
\alias{create_trans_dates}
\title{Calculates trans_dates data set for automated workflow}
\usage{
create_trans_dates(input_path_sst)
}
\arguments{
\item{input_path_sst}{Character string. Full path to the input data file}
}
\value{
ecodata::trans_dates data frame
}
\description{
This uses a static input file from Kevin Friedland
It is formatted exactly like the ecodata data object
}
\examples{
\dontrun{
# create the ecodata::trans_dates indicator for 2025
create_trans_dates(input_path_sst = "path/to/inputData.csv")

}


}
