% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-epu_strata.R
\docType{data}
\name{epu_strata}
\alias{epu_strata}
\title{Bottom Trawl Survey Strata and associated EPUs}
\format{
A data frame

\describe{
\item{EPU}{Ecological Production Unit}
\item{STRATUM}{STRATUM that comprise the EPU}

}
}
\source{
\url{https://noaa-edab.github.io/tech-doc/survdat.html}
}
\usage{
epu_strata
}
\description{
Ecological Production Units (EPUs) are defined (for Bottom Trawl Survey data) by
a set of strata in which the survey samples were taken
}
\keyword{datasets}
