% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_commercial_data.R
\name{get_commercial_data}
\alias{get_commercial_data}
\title{Get commercial data set to store somewhere}
\usage{
get_commercial_data(channel, outputPathDatasets = NULL)
}
\arguments{
\item{channel}{an Object inherited from DBIConnection-class.}

\item{outputPathDatasets}{Character string. Path to folder where data pull
should be saved (Default = NULL). If not NULL the pull will be saved to the
folder \code{outputPathDatasets} with names (\code{comdat.rds},\code{bennet.rds})}
}
\value{
A list of commercial data pulls
}
\description{
Pulls commercial data in multiple formats required for indicator generation
Menhaden data are removed from the commercial data pull since they are incomplete
}
\examples{
\dontrun{
channel <- dbutils::connect_to_database("server","user")
outputPathDataSets <- here::here()
rawData <- get_commercial_data(channel, outputPathDataSets)
}

}
