% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recreational_data.R
\name{get_recreational_data}
\alias{get_recreational_data}
\title{Pulls recreational HMS data from MRIP}
\usage{
get_recreational_data(output_path_indicators)
}
\arguments{
\item{output_path_indicators}{Character string. Full path the directory in which to store files.}
}
\value{
new.hms, a data frame with landings data from MRIP
}
\description{
This pulls landings data for HMS species in the North and Mid-Atlantic.
The function creates an rds and csv file for each species, to be stored in the designated output directory.
The function then combines all individual output files and saves as 'hms_mrip_' followed by the date in the output directory.
}
\examples{
\dontrun{
get_recreational_data(output_path_indicators = "path/to/output/directory/hms_mrip_date.csv")

}

}
