% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_data.R
\name{get_survey_data}
\alias{get_survey_data}
\title{Get survey data set to store somewhere}
\usage{
get_survey_data(channel, outputPathDatasets = NULL)
}
\arguments{
\item{channel}{an Object inherited from DBIConnection-class. This object is used to connect to communicate with the database engine.}

\item{outputPathDatasets}{Character string. Path to folder where data pull should be saved (Defaul = NULL).
If not NULL the pull will be saved to the folder \code{outputPathDatasets} with names (\code{lbatrossData.rds},\code{bigelowData.rds},\code{surveyNoLengths.rds})}
}
\value{
A list of survey data pulls
}
\description{
Pulls survey data in format required for indicator generation
}
