% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_status.R
\name{join_decoder}
\alias{join_decoder}
\title{Joins the stock status stock abbreviation lookup table}
\usage{
join_decoder(data, decoder)
}
\arguments{
\item{data}{the stock status data frame, typically from \code{stocksmart::stockAssessmentSummary}. Must have columns named \code{Entity.Name}, \code{Assessment.Year}, \code{F.Fmsy}, and \code{B.Bmsy}.}

\item{decoder}{a data frame that links the stock names with a code abbreviation to use in plotting, typically read from a CSV file. If set to \code{FALSE}, the function will return a data frame in a less processed form, like Sarah used to provide in the \code{assess.csv} file.}
}
\value{
a tibble
}
\description{
This function joins the stock status stock abbreviation lookup table.
}
