% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_status.R
\name{plot_stocksmart}
\alias{plot_stocksmart}
\title{Plot stock status data}
\usage{
plot_stocksmart(data, xmax, ymax)
}
\arguments{
\item{data}{the stock status data frame (processed with \code{prep_for_plotting()}}

\item{xmax}{the maximum value for the x-axis}

\item{ymax}{the maximum value for the y-axis}
}
\value{
a ggplot
}
\description{
This function creates a Kobe plot of stock status. Data must be prepared with \code{prep_for_plotting()}
}
