% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_status.R
\name{prep_for_plotting}
\alias{prep_for_plotting}
\title{Prepare stock status data for plotting}
\usage{
prep_for_plotting(data, council = c("MAFMC", "NEFMC"))
}
\arguments{
\item{data}{the stock status data frame}

\item{council}{a character vector of councils to filter by, default is c("MAFMC", "NEFMC"), which will return all stocks. Can specify only one council, e.g., "MAFMC" or "NEFMC", to filter to data for one council only.}
}
\value{
a tibble
}
\description{
This function wrangles the stock status data into a format suitable for plotting in the stock status vignette.
}
