% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comdat.R
\name{summarize_metrics}
\alias{summarize_metrics}
\title{Helper function for the summaries in get_comdat}
\usage{
summarize_metrics(data, value_col, metric_name, unit_name)
}
\arguments{
\item{data}{specify if all data, US only, seafood only should be used}

\item{metric_name}{metric to be summarized}

\item{unit_name}{description of what is being summarized and in what units}
}
\description{
Helper function for the summaries in get_comdat
}
