% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_biomass.r
\name{create_aggregate_biomass}
\alias{create_aggregate_biomass}
\title{Calculates aggregate_biomass data set for automated workflow}
\usage{
create_aggregate_biomass(inputPathSurvey, inputPathSpecies)
}
\arguments{
\item{inputPathSurvey}{Character string. Full path to the survdat data pull rds file}

\item{inputPathSpecies}{Character string. Full path to the species list data pull rds file}
}
\value{
ecodata::aggregate_biomass data frame
}
\description{
This uses the survdat data pull from the survey package and creates EPU
and shelfwide indicators. It is formatted exactly like the ecodata data object
}
\examples{
\dontrun{
# create the ecodata::aggregate_biomass indicator for 2025
create_aggregate_biomass(inputPathSurvey = "path/to/survdatData.rds"),
                         inputPathSpecies = "path/to/species.rds")

}


}
