% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comdat.R
\name{create_comdat}
\alias{create_comdat}
\title{Create data for ecodata::comdat}
\usage{
create_comdat(
  input_path_comdat,
  input_path_species,
  menhaden_path,
  outputPathDataSets
)
}
\arguments{
\item{input_path_comdat}{Character string. Path to commercial_comdat.rds}

\item{input_path_species}{Character string. Path to the 'SOE_species_list_24.RData' file.}

\item{menhaden_path}{Character string. Path to the menhaden data output by data-raw/create_menhaden_input.R}

\item{outputPathDataSets}{Character string. Path to folder where data pull should be saved}
}
\value{
list
\item{comdat}{\code{ecodata::comdat} data frame}
\item{comdat_species}{species data used to create the \code{comdat} indicator}
}
\description{
Processes and combines commercial landings and Menhaden data to produce a
summary of landings and revenue metrics for the State of the Ecosystem report.
}
