% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{create_condition}
\alias{create_condition}
\title{Create species condition indicator}
\usage{
create_condition(inputPath, inputPathLW, inputPathSpecies)
}
\arguments{
\item{inputPath}{Character string. Full path to the condition data pull rds file.}

\item{inputPathLW}{Character string. Full path to the LWparams csv file in 'EDAB_Resources/workflow_resources/soe_workflows'.}

\item{inputPathSpecies}{Character string. Full path to the species.codes csv file in 'EDAB_Resources/workflow_resources/soe_workflows'.}
}
\value{
condition, ecodata::condition data frame
}
\description{
This function calculates mean relative condition calculated from the NEFSC bottom trawl survey
Methods derived from Laurel Smith (https://github.com/Laurels1/Condition/blob/master/R/RelConditionEPU.R)
}
\examples{
\dontrun{
#create the ecodata::condition indicator
create_condition(
 inputPathLW = "path/to/LWparams.csv",
 inputPathSpecies = "path/to/species.codes.csv")

}

}
