% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatwave.R
\name{create_heatwave}
\alias{create_heatwave}
\title{Calculates the heatwave indicator}
\usage{
create_heatwave(
  inputPathGBBot,
  inputPathGOMBot,
  inputPathMABBot,
  inputPathGBSurf,
  inputPathGOMSurf,
  inputPathMABSurf
)
}
\arguments{
\item{inputPathGBBot}{Character string. Full path to the GB GLORYS input file from Joe Caracappa}

\item{inputPathGOMBot}{Character string. Full path to the GOM GLORYS input file from Joe Caracappa}

\item{inputPathMABBot}{Character string. Full path to the MAB GLORYS input file from Joe Caracappa}

\item{inputPathGBSurf}{Character string. Full path to the GB OISST input file from Kim Hyde}

\item{inputPathGOMSurf}{Character string. Full path to the GOM OISST input file from Kim Hyde}

\item{inputPathMABSurf}{Character string. Full path to the MAB OISST input file from Kim Hyde}
}
\value{
ecodata::heatwave data frame
}
\description{
Combines the surface SST, and the bottom temperature components
into a single indicator
}
\examples{
\dontrun{
}

}
