% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_inshore_survey.R
\name{create_mass_inshore_survey}
\alias{create_mass_inshore_survey}
\title{Function to process Massachusetts inshore bottom trawl survey for automated workflow}
\usage{
create_mass_inshore_survey(inputPathSurvey, inputPathSpecies)
}
\arguments{
\item{inputPathSpecies}{Character string. Full path to the species list data pull rds file}

\item{inputPathMassSurvey}{Character string. Full path to the mass inshore data pull rds file created by get_mass_survey}
}
\value{
ecodata::ma_inshore_survey data frame
}
\description{
Data include aggregated time series of inshore fishery-independent trawl survey data from Massachusetts waters.
MA inshore surveys have been performed biannually in the spring and fall since 1978.
}
\examples{
\dontrun{
# create the ecodata::ma_inshore_survey indicator
create_mass_inshore_survey(inputPathSurvey <- here::here("survdat.mass.rds"), 
                         inputPathSpecies <- "/home/<user>/EDAB_Datasets/SOE_species_list_24.rds")

}


}
