% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_hms.R
\name{create_rec_hms}
\alias{create_rec_hms}
\title{Creates recreational HMS indicator}
\usage{
create_rec_hms(inputPath, inputKey)
}
\arguments{
\item{inputPath}{Character string. Full path to the data from the 'pull_rec_hms' csv file in EDAB_Dev.}

\item{inputKey}{Character string. Full path to the hms_key file in EDAB_Resources ("EDAB_Resources/workflow_resources/soe_workflows/hms_key.csv")}
}
\value{
rec_hms, ecodata::rec_hms data frame
}
\description{
This uses the output file from 'pull_rec_hms' as an input parameter
The function formats the data to be used in the SOE
It is formatted exactly like the ecodata data object
}
\examples{
\dontrun{
#create the ecodata::rec_hms indicator
create_rec_hms(
 inputKey = "path/to/hms_key.csv")

}

}
