% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_dist.R
\name{create_species_dist}
\alias{create_species_dist}
\title{Species distribution on the NES.}
\usage{
create_species_dist(
  inputPathSurvey,
  inputPathSpecies,
  staticPath,
  outputPath = NULL
)
}
\arguments{
\item{inputPathSurvey}{Character string. Full path to the survdat data pull rds file}

\item{inputPathSpecies}{Character string. Full path to the species list data pull rds file}

\item{staticPath}{Character string. Path to folder for static files for depth and coast shape are saved}

\item{outputPath}{Character string. Path to folder where data pull should be saved}
}
\value{
ecodata::species_dist data frame
}
\description{
Data include time series of depth, distance from shelf and distance along shelf.
}
\examples{
\dontrun{
# create the ecodata::species_dist indicator
create_species_dist(inputPathSurvey <- here::here("surveyNoLengths.rds"),
                         inputPathSpecies <- "/home/<user>/EDAB_Datasets/SOE_species_list_24.rds",
                         outputPath <- here::here())

}


}
