% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_groupings.R
\name{create_species_groupings}
\alias{create_species_groupings}
\title{Calculates species_groupings data set for automated workflow}
\usage{
create_species_groupings(input_path_species_list, input_path_functional_group)
}
\arguments{
\item{input_path_species_list}{Character string. Full path to the original rds file from 2024}

\item{input_path_functional_group}{Character string. Full path to the species functional group csv file. SVSPP codes are mapped to functional group}
}
\value{
ecodata::species_groupings data frame
}
\description{
This data object is used in many of the other workflows in table joins
It is formatted exactly like the ecodata data object
}
\examples{
\dontrun{
# create the ecodata::species_groupings table
create_species_groupings(input_path_species_list = "path/to/SOE_species_list_old.rds",
                         input_path_functional_group = "path/to/functional_groups_list.csv")

}

}
