% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_commercial_data.R
\name{get_commercial_data}
\alias{get_commercial_data}
\title{Get commercial data set to store somewhere}
\usage{
get_commercial_data(channel)
}
\arguments{
\item{channel}{an Object inherited from DBIConnection-class.}
}
\value{
A list of commercial data pulls
}
\description{
Pulls commercial data in multiple formats required for indicator generation
Menhaden data are removed from the commercial data pull since they are incomplete
}
\examples{
\dontrun{
channel <- dbutils::connect_to_database("server","user")
rawData <- get_commercial_data(channel)
}

}
